Move template definitions from hdf5.cc to hdf5.h
